feat: basic test 2. 로그인/로그아웃 기능 추가
