+++ src/store/state.js
@@ -1,0 +1,3 @@
+export const state = {
+  loggedIn: !!localStorage.getItem("user"),
+};
